Allow default construction when GC'd
This addition makes it easier to use `SCM_convertible<GC_Protected>`
member variables in structure types that use default initialisation.

To facilitate this, this patch special-cases `SCM_UNDEFINED` as a
value that is never GC protected or unprotected.  For 
`SCM_UNDEFINED`, GC protection is not required because it is an
immediate value.  However, it must still be protected/unprotected in
pairs due to the internal implementation of `scm_gc_protect_object()`.